Refactored ActivityCategory entity and repository, added check by same-name-activity in editing and same-email-user in editing

+++ src/main/java/com/epam/billing/repository/AbstractRepository.java
@@ -22,5 +22,5 @@ public abstract class AbstractRepository<T> {
 
     public abstract boolean delete(T t);
 
-    public abstract boolean existById(long id);
+    public abstract boolean existById(int id);
 }
