+++ src/main/java/com/epam/billing/service/UserService.java
@@ -2,25 +2,18 @@ package com.epam.billing.service;
 
 import com.epam.billing.entity.User;
 import com.epam.billing.repository.UserRepository;
-
 import java.util.List;
 import java.util.Optional;
 
 public class UserService {
 
-
     private final UserRepository userRepository;
-
 
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
 
-
-
-    public List<User> getAll() {
-        return userRepository.getAll();
-    }
+    public List<User> getAll() { return userRepository.getAll(); }
 
     public void save(User user) {
         userRepository.save(user);
@@ -30,7 +23,7 @@ public class UserService {
         return userRepository.delete(user);
     }
 
-    public boolean existById(long id) {
+    public boolean existById(int id) {
         return userRepository.existById(id);
     }
